+++ sort.h
@@ -26,7 +26,7 @@ void selection_sort(int *, size_t);
 void selection_sort_swap(int *, int *);
 void quick_sort(int *array, size_t size);
 void quick_sort_sort(int *, size_t, size_t, size_t);
-void quick_sort_sort_swap(size_t *a, size_t *b);
+void quick_sort_sort_swap(int *a, int *b);
 
 /* Support Functions */
 void print_array(const int *array, size_t size);
